Fix ControllerUser after the two reverts
EJB exceptions are handled by the controller.

When creating a new user, it insert the tuple in the database. If
the mail if already used, it throwse an SQL exception, catch by the
EJB container, which throws an EJB excpetion catched by the
controllerUser.

The same mechanism takes place for the login, when no the mail is
not in the database.